--- src/daemon1.py
@@ -1,54 +1,0 @@
-#!/usr/bin/env python
-import argparse
-from multiprocessing import Process
-import logging
-from util import setup_logging
-
-from rq import Connection, Worker
-
-def start_worker(args: argparse.Namespace, process_index: int) -> None:
-    setup_logging()
-    logging.info(
-        "Agent [%s] running (process %s)...", args.group_id, process_index
-    )
-
-    with Connection():
-        w = Worker([args.group_id])
-        w.work()
-
-
-def main() -> None:
-    """Spawns a new agent process. Use argv if you want to use a different
-    group_id (it's `default` by default)
-    """
-
-    parser = argparse.ArgumentParser(description="Start mquery daemon.")
-    parser.add_argument(
-        "group_id",
-        help="Name of the agent group to join to",
-        nargs="?",
-        default="default",
-    )
-    parser.add_argument(
-        "--scale",
-        type=int,
-        help="Specifies the number of concurrent processes to use.",
-        default=1,
-    )
-
-    args = parser.parse_args()
-    if args.scale > 1:
-        children = [
-            Process(target=start_worker, args=(args, i))
-            for i in range(args.scale)
-        ]
-        for child in children:
-            child.start()
-        for child in children:
-            child.join()
-    else:
-        start_worker(args, 0)
-
-
-if __name__ == "__main__":
-    main()
